rename to upgrade_to_current_contract
