+++ crates/contract/tests/sandbox/mod.rs
@@ -1,9 +1,9 @@
 pub mod ckd;
-pub mod contract_upgrade;
 pub mod integration_tee_cleanup_after_resharing;
 pub mod sign;
 pub mod tee;
 pub mod updates;
+pub mod upgrade_to_current_contract;
 pub mod vote;
 
 pub mod common;
